% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roster.R
\name{add_student}
\alias{add_student}
\title{Add Student To Roster}
\usage{
add_student(Roster, Name, Experience = NA)
}
\arguments{
\item{Roster}{A data frame representing the current student roster. It should contain the columns: Name, Present, and Experience.}

\item{Name}{A character string indicating the students name. Must not already exist in the roster.}

\item{Experience}{A character string indicating the student's experience level. Must be one of the following: Beginner, Intermediate, or Advanced.}
}
\value{
A data frame representing the updated roster (Name and Experience level)
}
\description{
Adds a new student to the roster with their name and experience level.
}
\details{
The student is initially marked as absent. The updated roster is returned and sorted
alphabetically by name. Duplicate names are not allowed.
}
\examples{
Roster <- create_roster()
Roster <- add_student(Roster, "Nicola", "Beginner")
Roster <- add_student(Roster, "Keith", "Intermediate")
Roster <- add_student(Roster, "John", "Advanced")
}
