% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pairing.R
\name{create_pair_history}
\alias{create_pair_history}
\title{Initialise Empty Pairing History}
\usage{
create_pair_history()
}
\value{
An empty data frame with the columns: Class, Semester, Week, Student1, Student2.
}
\description{
Initialises an empty data frame to store historical pairings of students over classes, semesters, and weeks.
}
\details{
Initialises an empty pairing history. Essential for the start.
}
\examples{
history <- create_pair_history()
}
