% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roster.R
\name{create_roster}
\alias{create_roster}
\title{Initialise Empty Roster}
\usage{
create_roster()
}
\value{
An empty dataframe with columns Name (Character), Present (Logical - TRUE/FALSE), Experience (Factor with 3 levels: Beginner, Intermediate, Advanced)
}
\description{
Creates an empty roster for population into three columns: Name, Present, and Experience.
}
\examples{
Roster <- create_roster()
Roster
}
