% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pairing.R
\name{generate_all_pairs}
\alias{generate_all_pairs}
\title{Generate All Unique Student Pairs}
\usage{
generate_all_pairs(present_names)
}
\arguments{
\item{present_names}{A character vector of student names currently present in class.}
}
\value{
A data frame with columns Student1 and Student2.
}
\description{
Generates all possible unique pairs (combinations of two) from a list of student names.
}
\details{
This is a helper function. Core logic to create all possible student pairs, which is then used by generate_session_pairs()
}
\examples{
generate_all_pairs(present_students)
}
