% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roster.R
\name{mark_all_absent}
\alias{mark_all_absent}
\title{Mark All Students Absent}
\usage{
mark_all_absent(Roster)
}
\arguments{
\item{Roster}{A data frame representing the current student roster. Must contain a Present column.}
}
\value{
A data frame with all students marked as absent.
}
\description{
Sets present status to FALSE for all students in roster.
}
\details{
Useful at start of class when needing to reset roster.
}
\examples{
Roster <- create_roster()
Roster <- add_student(Roster, "Nicola", "Beginner")
Roster <- add_student(Roster, "Keith", "Intermediate")
Roster <- add_student(Roster, "John", "Advanced")
Roster <- mark_all_absent(Roster)
}
