% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roster.R
\name{mark_present}
\alias{mark_present}
\title{Mark Student As Present}
\usage{
mark_present(Roster, Names)
}
\arguments{
\item{Roster}{A data frame representing the current student roster. Must contain columns Name and Present.}

\item{Names}{A character vector of student names to be marked present.}
}
\value{
A data frame representing the updated roster with students marked as present.
}
\description{
Sets present status to TRUE for list of students by name.
}
\examples{
Roster <- create_roster()
Roster <- add_student(Roster, "Nicola", "Beginner")
Roster <- add_student(Roster, "Keith", "Intermediate")
Roster <- add_student(Roster, "John", "Advanced")
Roster <- mark_present(roster, c("Nicola", "Keith"))
}
