% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Roster.R
\name{remove_student}
\alias{remove_student}
\title{Remove Student From Roster}
\usage{
remove_student(Roster, Name)
}
\arguments{
\item{Roster}{A data frame representing the current student roster. It should contain the column Name.}

\item{Name}{A character string specifying the name of the student to remove.}
}
\value{
A data frame representing the updated roster, with the specified student removed.
}
\description{
Removes a student by name from the roster.
}
\examples{
Roster <- create_roster()
Roster <- add_student(Roster, "Nicola", "Beginner")
Roster <- add_student(Roster, "Keith", "Intermediate")
Roster <- add_student(Roster, "John", "Advanced")
Roster <- remove_student(Roster, "Nicola") # dropped out
}
